+++ PSJobLogger.build.ps1
@@ -1,6 +1,7 @@
 #requires -modules InvokeBuild
 task Clean {
     Remove-Item PSJobLogger/PSJobLogger.psd1 -Force -ErrorAction SilentlyContinue
+    Remove-Item hack/test.log -Force -ErrorAction SilentlyContinue
 }
 
 task Lint {
@@ -8,14 +9,14 @@ task Lint {
 }
 
 task Test {
-    Remove-Module PSJobLogger -Force -ErrorAction Ignore
-    Import-Module ./PSJobLogger/PSJobLogger.psd1 -Force -ErrorAction Stop
+    Remove-Module PSJobLogger -Force -ErrorAction SilentlyContinue
+    Import-Module ./PSJobLogger -Force
     Invoke-Pester
 }
 
 task Install-Dependencies {
-    @('Pester','PSScriptAnalyzer') | ForEach-Object {
-        Install-Module $_ -Force -ErrorAction Stop
+    foreach ($module in 'Pester','PSScriptAnalyzer') {
+        Install-Module $module -Scope CurrentUser -Force
     }
 }
 
@@ -27,9 +28,9 @@ task Build-Manifest {
         Copyright = '(c) 2023 Alexander W Lew. All Rights Reserved.'
         CompanyName = 'Alan Lew'
         RootModule = 'PSJobLogger.psm1'
-        ModuleVersion = '0.4.0'
+        ModuleVersion = '0.5.0'
         Description = 'A logging class suitable for use with ForEach-Object -Parallel -AsJob'
-        PowerShellVersion = '5.0'
+        PowerShellVersion = '5.1'
         NestedModules = @(
             'DictLogger.psm1'
         )
@@ -61,10 +62,12 @@ task Build-Manifest {
             'PSJobLoggerStreamVerbose',
             'PSJobLoggerStreamDebug',
             'PSJobLoggerStreamInformation',
+            'PSJobLoggerStreamHost',
             'PSJobLoggerStreamProgress',
             'PSJobLoggerLogStreams'
+            'PSJobLoggerPlainTextLogStreams'
         )
-        FileList = 'PSJobLogger.psd1','PSJobLogger.psm1', 'DictLogger.psm1'
+        FileList = 'PSJobLogger.psd1','PSJobLogger.psm1', 'DictLogger.psm1', 'en-US/about_PSJobLogger.help.txt'
         Tags = 'ForEach-Object','Parallel','AsJob','Logging','PSEdition_Core','Windows','Linux','MacOS'
         ProjectUri = 'https://github.com/neflyte/PSJobLogger'
         LicenseUri = 'https://github.com/neflyte/PSJobLogger/blob/main/LICENSE'
@@ -74,11 +77,11 @@ task Build-Manifest {
 }
 
 task Mp3test {
+    Remove-Module PSJobLogger -Force -ErrorAction SilentlyContinue
+    Import-Module ./PSJobLogger -Force
     Push-Location hack
     try {
         Remove-Item test.log -Force -ErrorAction SilentlyContinue
-        Remove-Module PSJobLogger -Force -ErrorAction SilentlyContinue
-        Import-Module ../PSJobLogger -Force -ErrorAction Stop
         ./Process-Mp3Files.ps1 -Directory $HOME/Music/share -Logfile test.log
     } finally {
         Pop-Location