+++ Tests/DictLogger.Tests.ps1
@@ -1,22 +1,24 @@
+using module ../PSJobLogger
 using namespace System.Collections
 using namespace System.Collections.Concurrent
 
 InModuleScope PSJobLogger {
+    BeforeAll {
+        $LoggerName = 'DictLogger-test'
+    }
     Describe 'DictLogger' {
-        BeforeAll {
-            [Diagnostics.CodeAnalysis.SuppressMessageAttribute('UseDeclaredVarsMoreThanAssignments', '')]
-            $loggerName = 'DictLogger-test'
-        }
-    
         BeforeEach {
-            [Diagnostics.CodeAnalysis.SuppressMessageAttribute('UseDeclaredVarsMoreThanAssignments', '')]
-            $logger = Initialize-PSJobLoggerDict -Name:$loggerName -LogFile:'' -UseQueues -ProgressParentId:-1
+            $logger = Initialize-PSJobLoggerDict -Name $LoggerName -LogFile '' -UseQueues -ProgressParentId -1
         }
-    
+
         Context 'Initialize-PSJobLoggerDict' {
             It 'initializes correctly' {
                 $logger | Should -Not -BeNullOrEmpty
                 $logger.Name | Should -BeExactly $loggerName
+                $logger.Logfile | Should -BeExactly ''
+                $logger.ShouldLogToFile | Should -BeFalse
+                $logger.VerbosePref | Should -BeExactly 'SilentlyContinue'
+                $logger.DebugPref | Should -BeExactly 'SilentlyContinue'
                 $logger.UseQueues | Should -BeTrue
                 $logger.ProgressParentId | Should -BeExactly -1
                 $logger.Streams | Should -Not -BeNullOrEmpty
@@ -35,7 +37,7 @@ InModuleScope PSJobLogger {
                 }
             }
         }
-    
+
         Context 'Write-LogProgress' {
             It 'adds a new map' {
                 Write-LogProgress -LogDict $logger -Id 'foo' -ArgumentMap @{ Id = 1; Activity = 'bar' }
@@ -88,10 +90,10 @@ InModuleScope PSJobLogger {
                 $progressArgs.Completed | Should -BeTrue
             }
         }
-    
+
         Context 'Write-LogOutput' {
             It 'enqueues a message' {
-                [ConcurrentQueue[String]]$successTable = $logger.Streams.$PSJobLoggerStreamSuccess
+                $successTable = $logger.Streams.$PSJobLoggerStreamSuccess
                 $successTable.Count | Should -BeExactly 0
                 Write-LogOutput -LogDict $logger -Message 'foo'
                 $successTable = $logger.Streams.$PSJobLoggerStreamSuccess
@@ -99,6 +101,46 @@ InModuleScope PSJobLogger {
                 $successTable[0] | Should -Contain 'foo'
             }
         }
-    
+
+        Context 'Set-Logfile' {
+            BeforeEach {
+                $logfile = New-TemporaryFile
+            }
+            AfterEach {
+                Remove-Item $logfile -Force
+            }
+            It 'sets the correct dict state' {
+                Set-Logfile -LogDict $logger -Filename $logfile
+                $logger.Logfile | Should -BeExactly $logfile.ToString()
+                $logger.ShouldLogToFile | Should -BeTrue
+            }
+        }
+
+        Context 'Write-MessageToLogFile' {
+            BeforeEach {
+                $logfile = New-TemporaryFile
+            }
+            AfterEach {
+                Remove-Item $logfile -Force
+            }
+            It 'writes to the log file' {
+                Set-Logfile -LogDict $logger -Filename $logfile
+                Write-MessageToLogfile -LogDict $logger -Stream $PSJobLoggerStreamSuccess -Message '1: LOG SUCCESS'
+                Write-MessageToLogfile -LogDict $logger -Stream $PSJobLoggerStreamError -Message '2: LOG ERROR'
+                Write-MessageToLogfile -LogDict $logger -Stream $PSJobLoggerStreamWarning -Message '3: LOG WARNING'
+                Write-MessageToLogfile -LogDict $logger -Stream $PSJobLoggerStreamVerbose -Message '4: LOG VERBOSE'
+                Write-MessageToLogfile -LogDict $logger -Stream $PSJobLoggerStreamDebug -Message '5: LOG DEBUG'
+                Write-MessageToLogfile -LogDict $logger -Stream $PSJobLoggerStreamInformation -Message '6: LOG INFO'
+                Write-MessageToLogfile -LogDict $logger -Stream $PSJobLoggerStreamHost -Message '6: LOG HOST'
+                $logContents = Get-Content $logfile -Raw
+                $logContents | Should -Match '1: LOG SUCCESS'
+                $logContents | Should -Match '2: LOG ERROR'
+                $logContents | Should -Match '3: LOG WARNING'
+                $logContents | Should -Match '4: LOG VERBOSE'
+                $logContents | Should -Match '5: LOG DEBUG'
+                $logContents | Should -Match '6: LOG INFO'
+                $logContents | Should -Match '6: LOG HOST'
+            }
+        }
     }
 }