+++ Tests/PSJobLogger.Tests.ps1
@@ -1,15 +1,16 @@
-using module '../PSJobLogger/PSJobLogger.psd1'
+using module ../PSJobLogger
 using namespace System.Collections.Concurrent
+using namespace System.Collections.Generic
+using namespace System.IO
+Import-Module (Join-Path $PSScriptRoot 'Helpers.psm1') -Force
 
 InModuleScope PSJobLogger {
-    Describe 'PSJobLogger' {
-        BeforeAll {
-            [Diagnostics.CodeAnalysis.SuppressMessageAttribute('UseDeclaredVarsMoreThanAssignments', '')]
-            $LoggerName = 'PSJobLogger-test'
-        }
+    BeforeAll {
+        $LoggerName = 'PSJobLogger-test'
+    }
 
+    Describe 'PSJobLogger' {
         BeforeEach {
-            [Diagnostics.CodeAnalysis.SuppressMessageAttribute('UseDeclaredVarsMoreThanAssignments', '')]
             $logger = [PSJobLogger]::new($LoggerName, '', $true, -1)
         }
 
@@ -18,6 +19,10 @@ InModuleScope PSJobLogger {
                 $logger | Should -Not -BeNullOrEmpty
                 $logger.Name | Should -BeExactly $LoggerName
                 $logger.UseQueues | Should -BeTrue
+                $logger.Logfile | Should -BeExactly ''
+                $logger.ShouldLogToFile | Should -BeFalse
+                $logger.VerbosePref | Should -BeExactly 'SilentlyContinue'
+                $logger.DebugPref | Should -BeExactly 'SilentlyContinue'
                 $logger.Streams.Keys.Count | Should -Be $([PSJobLogger]::LogStreams.Keys).Count
                 $logger.Streams.Keys | ForEach-Object {
                     switch ($_) {
@@ -106,11 +111,70 @@ InModuleScope PSJobLogger {
             }
         }
 
+        Context 'FlushMessages' {
+            BeforeEach {
+                $logger.VerbosePref = 'Continue'
+                $logger.DebugPref = 'Continue'
+                $logCapture = New-TemporaryFile
+            }
+
+            AfterEach {
+                Remove-Item $logCapture -Force
+            }
+
+            It 'flushes to plain text streams' {
+                $logger.Output('1: LOG SUCCESS')
+                $logger.Error('2: LOG ERROR')
+                $logger.Warning('3: LOG WARNING')
+                $logger.Verbose('4: LOG VERBOSE')
+                $logger.Debug('5: LOG DEBUG')
+                $logger.Information('6: LOG INFO')
+                $logger.Host('6: LOG HOST')
+                FlushAndCapture -JobLogger $logger -LogCapture $logCapture
+                $captured = Get-Content $logCapture -Raw
+                # $captured -match '1: LOG SUCCESS' | Should -BeTrue
+                $captured -match '2: LOG ERROR' | Should -BeTrue
+                $captured -match '3: LOG WARNING' | Should -BeTrue
+                $captured -match '4: LOG VERBOSE' | Should -BeTrue
+                $captured -match '5: LOG DEBUG' | Should -BeTrue
+                $captured -match '6: LOG INFO' | Should -BeTrue
+                $captured -match '6: LOG HOST' | Should -BeTrue
+            }
+        }
+
+        Context 'Logfile' {
+            BeforeEach {
+                $logfile = New-TemporaryFile
+            }
+            AfterEach {
+                Remove-Item $logfile -Force
+            }
+            It 'Writes to a log file' {
+                $logger.SetLogfile($logfile)
+                $logger.Output('1: LOG SUCCESS')
+                $logger.Error('2: LOG ERROR')
+                $logger.Warning('3: LOG WARNING')
+                $logger.Verbose('4: LOG VERBOSE')
+                $logger.Debug('5: LOG DEBUG')
+                $logger.Information('6: LOG INFO')
+                $logger.Host('6: LOG HOST')
+                $logger.FlushPlainTextStreams()
+                $logfileContents = Get-Content $logfile -Raw
+                $logfileContents -match '1: LOG SUCCESS' | Should -BeTrue
+                $logfileContents -match '2: LOG ERROR' | Should -BeTrue
+                $logfileContents -match '3: LOG WARNING' | Should -BeTrue
+                $logfileContents -match '4: LOG VERBOSE' | Should -BeTrue
+                $logfileContents -match '5: LOG DEBUG' | Should -BeTrue
+                $logfileContents -match '6: LOG INFO' | Should -BeTrue
+                $logfileContents -match '6: LOG HOST' | Should -BeTrue
+            }
+        }
+
         Context 'asDictLogger' {
             It 'converts from a class' {
                 $dictLogger = $logger.asDictLogger()
                 $dictLogger | Should -Not -BeNullOrEmpty
-                $expectedKeys = @('Name', 'Logfile', 'UseQueues', 'ProgressParentId', 'Streams')
+                $expectedKeys = 'Name', 'Logfile', 'ShouldLogToFile', 'VerbosePref', 'DebugPref', 'UseQueues', 'ProgressParentId', 'Streams'
                 foreach ($key in $expectedKeys) {
                     $dictLogger.ContainsKey($key) | Should -BeTrue
                     $dictLogger.$key | Should -Not -Be $null
@@ -120,10 +184,14 @@ InModuleScope PSJobLogger {
 
         Context 'Initialize-PSJobLogger' {
             It 'initializes' {
-                $jobLogger = Initialize-PSJobLogger -Name $LoggerName -Logfile '' -UseQueues -ParentProgressId -1
+                $jobLogger = Initialize-PSJobLogger -Name $LoggerName -Logfile '' -UseQueues -ProgressParentId -1
                 $jobLogger | Should -Not -BeNullOrEmpty
                 $jobLogger.Name | Should -BeExactly $LoggerName
                 $jobLogger.UseQueues | Should -BeTrue
+                $jobLogger.Logfile | Should -BeExactly ''
+                $jobLogger.ShouldLogToFile | Should -BeFalse
+                $logger.VerbosePref | Should -BeExactly 'SilentlyContinue'
+                $logger.DebugPref | Should -BeExactly 'SilentlyContinue'
                 $jobLogger.Streams.Keys.Count | Should -Be $([PSJobLogger]::LogStreams.Keys).Count
                 $jobLogger.Streams.Keys | ForEach-Object {
                     switch ($_) {