+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2023-12-30
+### Added
+- Method to flush Plain Text streams (i.e. all streams except Progress)
+- Unit tests to cover writing log messages to streams and to a file
+- Function to set the logfile in the DictLogger
+
+### Changed
+
+### Removed
+
 ## [0.4.0] - 2023-09-03
 ### Added
 - Methods to convert a DictLogger to a PSJobLogger class and back again